+++ src/Headless.Payments.Paymob.CashIn/Models/Payment/CashInWalletPayResponse.cs
@@ -9,7 +9,7 @@ namespace Headless.Payments.Paymob.CashIn.Models.Payment;
 public sealed class CashInWalletPayResponse
 {
     [JsonPropertyName("id")]
-    public int Id { get; init; }
+    public long Id { get; init; }
 
     [JsonPropertyName("pending")]
     public bool Pending { get; init; }
@@ -39,10 +39,10 @@ public sealed class CashInWalletPayResponse
     public bool Is3DSecure { get; init; }
 
     [JsonPropertyName("integration_id")]
-    public int IntegrationId { get; init; }
+    public long IntegrationId { get; init; }
 
     [JsonPropertyName("profile_id")]
-    public int ProfileId { get; init; }
+    public long ProfileId { get; init; }
 
     [JsonPropertyName("has_parent_transaction")]
     public bool HasParentTransaction { get; init; }
@@ -76,7 +76,7 @@ public sealed class CashInWalletPayResponse
     public long RefundedAmountCents { get; init; }
 
     [JsonPropertyName("source_id")]
-    public int SourceId { get; init; }
+    public long SourceId { get; init; }
 
     [JsonPropertyName("is_captured")]
     public bool IsCaptured { get; init; }
@@ -85,7 +85,7 @@ public sealed class CashInWalletPayResponse
     public long CapturedAmount { get; init; }
 
     [JsonPropertyName("owner")]
-    public int Owner { get; init; }
+    public long Owner { get; init; }
 
     [JsonPropertyName("error_occured")]
     public bool ErrorOccured { get; init; }
